requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# File processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
pypandoc>=1.12

# Cloud storage
boto3>=1.29.0
azure-storage-blob>=12.19.0

# LLM providers
openai>=1.3.0
anthropic>=0.7.0
ollama>=0.1.0

# Utilities
aiofiles>=23.2.1
asyncio-throttle>=1.0.2
httpx>=0.25.0

# Logging and monitoring
structlog>=23.2.0

# Type checking (optional)
typing-extensions>=4.8.0

